+++ CHANGELOG.md
@@ -4,14 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Improve getting started documentation
+
 ## [0.5.1] - 2022-02-08
 
 ### Fixed
-* `.dockerignore` not correctly included in package data.
-* Setup documentation URL.
+- `.dockerignore` not correctly included in package data.
+- Setup documentation URL.
 
 ### Changed
-* Auto-redirect for docs index.
+- Auto-redirect for docs index.
 
 ## [0.5.0] - 2022-02-08