+++ CHANGELOG.md
@@ -6,8 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Missing files in git for example projects.
+
 ### Changed
-- Improve getting started documentation
+- Improve getting started documentation.
 
 ## [0.5.1] - 2022-02-08