+++ CHANGELOG.md
@@ -4,13 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.6.0 - Unreleased]
 
 ### Added
 - `FileReferenceCacher` for storing lists of referenced file paths without keeping
   file contents in memory.
 - Automatic reportable caching. Reportables of Stages that short-circuit will now
   reload and display in the report.
+- Misc example project folder for experiments demonstrating various curifactory
+  features.
 
 ### Changed
 - Improve getting started documentation.